+++ source/docs/guide/src/SUMMARY.md
@@ -40,6 +40,7 @@
 - [Developing proofs](develop_proofs.md)
     - [Using assert and assume](assert_assume.md)
     - [Devising loop invariants](invariants.md)
+    - [Proving absence of overflow](overflow.md)
 - [Quantifiers](quants.md)
     - [forall and triggers](forall.md)
     - [Multiple variables, multiple triggers, matching loops](multitriggers.md)
